requirements.txt
Django==4.2.7
channels==4.0.0
channels-rabbitmq==4.0.1
django-notifs==4.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
# psycopg2-binary==2.9.9
python-decouple==3.8 